requirements.txt
streamlit>=1.37.1
python-dotenv>=1.0.1
requests>=2.32.0
jira>=3.8.0
